refact(fock/pure): Extract passive linear gates
The passive linear gates got extracted, because `calculations.py` got too
large.
